feat()：ijkplayer的实例代码抽到一个activity里
