+++ app/src/main/java/ryan/media/iavideo/MainActivity.java
@@ -1,21 +1,17 @@
 package ryan.media.iavideo;
 
-import androidx.appcompat.app.AppCompatActivity;
-
+import android.content.Intent;
 import android.os.Bundle;
-import android.os.Environment;
-import android.util.Log;
-import android.view.MotionEvent;
-import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import java.io.File;
-import java.io.IOException;
+import androidx.activity.result.ActivityResult;
+import androidx.activity.result.ActivityResultCallback;
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
+import androidx.appcompat.app.AppCompatActivity;
 
 import ryan.media.iavideo.databinding.ActivityMainBinding;
 import ryan.utils.PermissionUtil;
-import ryan.utils.Constant;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -25,6 +21,13 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private ActivityMainBinding binding;
+    private ActivityResultLauncher launcher =
+            registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), new ActivityResultCallback<ActivityResult>() {
+                @Override
+                public void onActivityResult(ActivityResult result) {
+
+                }
+            });
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -39,30 +42,6 @@ public class MainActivity extends AppCompatActivity {
         TextView tv = binding.sampleText;
         tv.setText(stringFromJNI());
         tv.bringToFront();
-
-        VideoPlayer ijkVideo = binding.ijkVideo;
-//        ijkVideo.setVideoListener(this);
-        String path = "/sdcard/DCIM/Camera";
-        String fileName = "DCIM/Camera/share_f032ca40160c3d22598faf07728e8f51.mp4";
-        String input = new File(
-                Environment.getExternalStorageDirectory(),
-                fileName
-        ).getAbsolutePath();
-        ijkVideo.setPath(input);
-        try {
-            ijkVideo.load();
-        } catch (IOException e) {
-            Toast.makeText(this,"播放失败",Toast.LENGTH_SHORT);
-            e.printStackTrace();
-        }
-        tv.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.e(Constant.TAG, "askjdh");
-                Toast.makeText(MainActivity.this,"开始",Toast.LENGTH_SHORT);
-                ijkVideo.start();
-            }
-        });
     }
 
     /**
@@ -70,4 +49,12 @@ public class MainActivity extends AppCompatActivity {
      * which is packaged with this application.
      */
     public native String stringFromJNI();
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        Intent intent = new Intent();
+        intent.setAction("ijk_video_player");
+        launcher.launch(intent);
+    }
 }
